added cranberry bushes, clean up a bunch of bad code

+++ src/main/java/net/dries007/tfc/common/blocks/berry_bush/DeadBerryBushBlock.java
@@ -7,48 +7,74 @@ import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.fluid.Fluid;
+import net.minecraft.fluid.FluidState;
+import net.minecraft.fluid.Fluids;
 import net.minecraft.state.StateContainer;
 import net.minecraft.util.ActionResultType;
+import net.minecraft.util.Direction;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockRayTraceResult;
+import net.minecraft.world.IWorld;
 import net.minecraft.world.World;
 import net.minecraft.world.server.ServerWorld;
 
 import net.dries007.tfc.common.TFCTags;
 import net.dries007.tfc.common.blocks.ForgeBlockProperties;
+import net.dries007.tfc.common.blocks.TFCBlockStateProperties;
+import net.dries007.tfc.common.fluids.FluidProperty;
+import net.dries007.tfc.common.fluids.IFluidLoggable;
 import net.dries007.tfc.common.tileentity.TickCounterTileEntity;
 import net.dries007.tfc.util.Helpers;
 import net.dries007.tfc.util.calendar.ICalendar;
 
-public class DeadBerryBushBlock extends SpreadingBushBlock
+public class DeadBerryBushBlock extends AbstractBerryBushBlock implements IFluidLoggable
 {
+    private static final FluidProperty FLUID = TFCBlockStateProperties.FRESH_WATER;
+
     public DeadBerryBushBlock(ForgeBlockProperties properties)
     {
-        super(properties, BerryBush.NOOP, () -> Blocks.AIR);
+        super(properties, BerryBush.NOOP);
+        registerDefaultState(getStateDefinition().any().setValue(getFluidProperty(), getFluidProperty().keyFor(Fluids.EMPTY)).setValue(STAGE, 0));
+    }
+
+    @Override
+    public BlockState updateShape(BlockState stateIn, Direction facing, BlockState facingState, IWorld worldIn, BlockPos currentPos, BlockPos facingPos)
+    {
+        if (!stateIn.canSurvive(worldIn, currentPos))
+        {
+            return Blocks.AIR.defaultBlockState();
+        }
+        else
+        {
+            final Fluid containedFluid = stateIn.getValue(getFluidProperty()).getFluid();
+            if (containedFluid != Fluids.EMPTY)
+            {
+                worldIn.getLiquidTicks().scheduleTick(currentPos, containedFluid, containedFluid.getTickDelay(worldIn));
+            }
+            return super.updateShape(stateIn, facing, facingState, worldIn, currentPos, facingPos);
+        }
     }
 
     @Override
     protected void createBlockStateDefinition(StateContainer.Builder<Block, BlockState> builder)
     {
-        builder.add(STAGE);
+        builder.add(getFluidProperty(), STAGE);
     }
 
     @Override
     public void randomTick(BlockState state, ServerWorld world, BlockPos pos, Random random)
     {
-        if (random.nextInt(15) == 0 && world.isEmptyBlock(pos.above()))
+        if (random.nextInt(15) == 0 && !world.getBlockState(pos.above()).is(TFCTags.Blocks.ANY_SPREADING_BUSH))
         {
             TickCounterTileEntity te = Helpers.getTileEntity(world, pos, TickCounterTileEntity.class);
             if (te != null)
             {
                 if (te.getTicksSinceUpdate() > ICalendar.TICKS_IN_DAY * 80)
                 {
-                    if (!world.getBlockState(pos.above()).is(TFCTags.Blocks.ANY_SPREADING_BUSH))
-                    {
-                        te.setRemoved();
-                        world.destroyBlock(pos, true);
-                    }
+                    te.setRemoved();
+                    world.destroyBlock(pos, true);
                 }
             }
         }
@@ -58,5 +84,18 @@ public class DeadBerryBushBlock extends SpreadingBushBlock
     public ActionResultType use(BlockState state, World worldIn, BlockPos pos, PlayerEntity player, Hand handIn, BlockRayTraceResult hit)
     {
         return ActionResultType.FAIL;
+    }
+
+    @Override
+    @SuppressWarnings("deprecation")
+    public FluidState getFluidState(BlockState state)
+    {
+        return IFluidLoggable.super.getFluidState(state);
+    }
+
+    @Override
+    public FluidProperty getFluidProperty()
+    {
+        return FLUID;
     }
 }
